fix: Some repository sample type-related issues feat: Add explicit ordering for some SQL queries fix: Split out task distribution queries, as specifying the extract keyword as a variable is not allowed it seems fix: Make sure to not count any records where success happens to be NULL (excluding for the queries where we group by success)

+++ src/main/java/com/docshifter/core/metrics/samples/TasksDistributionSample.java
@@ -5,6 +5,6 @@ package com.docshifter.core.metrics.samples;
  * @author Created by Juan Marques on 03/08/2021
  */
 public interface TasksDistributionSample {
-    byte getDateValue();
-    long getCount();
+    short getDateValue();
+    Long getCount();
 }
